notes/serializers.py
from rest_framework import serializers
from .models import Note

class NoteSerializers(serializers.ModelSerializer):
    class Meta:
        model = Note
        fields = '__all__'
        read_only_fields = ['category']
