requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
joblib==1.5.0
numpy==2.2.5
scikit-learn==1.6.1
scipy==1.15.3
sqlparse==0.5.3
threadpoolctl==3.6.0
drf-yasg==1.21.10
